split Fulfilment link handling behaviour out of EconomicEvent handler module
